Implement "Active Only" items filter
Hides all items which are either done or cancelled. Makes it easier
to track which items are to be taken care of.

There could be some more clear ways to have visual separation between
active and completed items, but this way is easy to implement and will
be handy anyway. Other ideas could be to implement some sort of group,
so that completed items are grouped in a nested node or at the back of
the list.